requirements.txt
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.7
gunicorn==21.2.0
python-dotenv==1.0.0
sendgrid==6.10.0
twilio==8.5.0
PyJWT==2.8.0
